+++ SimpleBrowser-P4.xcodeproj/xcuserdata/lukasznycz.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   uuid = "84C7B1C3-CACF-4D73-B4AD-F7FD328700F3"
+   type = "1"
+   version = "2.0">
+</Bucket>
